// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  VOZAC: 'VOZAC'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Plan = {
  STARTER: 'STARTER',
  PRO: 'PRO',
  ENTERPRISE: 'ENTERPRISE'
} as const

export type Plan = (typeof Plan)[keyof typeof Plan]


export const VrstaGuma = {
  LJETNE: 'LJETNE',
  ZIMSKE: 'ZIMSKE'
} as const

export type VrstaGuma = (typeof VrstaGuma)[keyof typeof VrstaGuma]